Modification de plusieurs routes: 	- la route pour la clé d'api peut maintenant générer des clés (UUID) et des comptes selon le besoin. 	- les routes ayant besoins d'une clé api ont maintenant cette validation sauf la route /type (prévoit de l'ajouter prochainement)

+++ src/Domain/Fakemon/Repository/FakemonRepository.php
@@ -25,6 +25,27 @@ class FakemonRepository
     }
 
     /**
+     * Permet de générer un uuid.
+     * Source: https://www.uuidgenerator.net/dev-corner/php
+     * 
+     * @return string un uuid.
+     */
+    private function guidv4($data = null)
+    {
+        // Generate 16 bytes (128 bits) of random data or use the data passed into the function.
+        $data = $data ?? random_bytes(16);
+        assert(strlen($data) == 16);
+
+        // Set version to 0100
+        $data[6] = chr(ord($data[6]) & 0x0f | 0x40);
+        // Set bits 6-7 to 10
+        $data[8] = chr(ord($data[8]) & 0x3f | 0x80);
+
+        // Output the 36 character UUID.
+        return vsprintf('%s%s-%s-%s-%s-%s%s%s', str_split(bin2hex($data), 4));
+    }
+
+    /**
      * Sélectionne l'api du compte fourni
      * 
      * @param array les informations du compte
@@ -37,24 +58,112 @@ class FakemonRepository
         $params = [
             "username" => $compte["username"] ?? ""
         ];
+        $query = $this->connection->prepare($sql);
+        $query->execute($params);
+
+        $result = $query->fetchAll(PDO::FETCH_ASSOC);
+
+        $apikey = $result[0] ?? [];
+
+        if (!empty($apikey)) {
+            if ($apikey["api_key"] == "") {
+                $resultat = $this->creerCleCompte($compte);
+                $apikey["api_key"] = $resultat["api_key"];
+            }
+        } else {
+            $apikey = [];
+        }
+
+        return $apikey;
+    }
+
+    /**
+     * Créer une clé api à un compte.
+     * 
+     * @param array les informations du compte.
+     * @return DataResponse
+     */
+    public function creerCleCompte(array $compte): array
+    {
+        $cleNouv = $this->guidv4();
+
+        $sql = "UPDATE usager SET api_key = :apikey WHERE username = :username ;";
+
+        $params = [
+            "apikey" => $cleNouv,
+            "username" => $compte["username"]
+        ];
+
+        $query = $this->connection->prepare($sql);
+        $query->execute($params);
+
+        $result = [
+            "api_key" => $cleNouv
+        ];
+
+
+        return $result;
+    }
+
+    /**
+     * Permet d'inserer un nouvel usager.
+     * 
+     * @param array les informations de l'usager
+     * @return DataResponse
+     */
+    public function creerUsager($compte): array
+    {
+        $sql = "INSERT INTO usager (username,password,api_key) 
+        VALUES (:username,:password,:api_key);";
+
+        $params = [
+            "username" => $compte["username"],
+            "password" => password_hash($compte["password"], PASSWORD_DEFAULT),
+            "api_key" => $this->guidv4()
+        ];
+
+        $query = $this->connection->prepare($sql);
+        $query->execute($params);
+
+        $idUsager = $this->connection->lastInsertId();
+        $resultat = $this->afficherUsagerById($idUsager);
+
+        return $resultat ?? [];
+    }
+
+
+    /**
+     * Sélectionne l'usager en question.
+     * 
+     * @param int l'id de l'usager
+     * @return DataResponse
+     */
+    public function afficherUsagerById(int $id): array
+    {
+        $sql = "SELECT * FROM usager WHERE id = :id ;";
+
+        $params = [
+            "id" => $id
+        ];
 
         $query = $this->connection->prepare($sql);
         $query->execute($params);
 
         $result = $query->fetchAll(PDO::FETCH_ASSOC);
 
-        
-        
+
+
         return $result[0] ?? [];
     }
 
-     /**
+
+    /**
      * Sélectionne le fakemon de l'id en question
      * 
-     * @param array les informations du compte
+     * @param int l'id du fakemon
      * @return DataResponse
      */
-    public function afficherFakemonById($id): array
+    public function afficherFakemonById(int $id): array
     {
         $sql = "SELECT * FROM creature WHERE id = :id ;";
 
@@ -67,33 +176,42 @@ class FakemonRepository
 
         $result = $query->fetchAll(PDO::FETCH_ASSOC);
 
-        
-        
+
+
         return $result[0] ?? [];
     }
 
     /**
      * Permet d'insert un nouveau fakemon
      * 
-     * @param array les informations du fakemon
+     * @param array les informations du fakemon.
+     * @param string l'username du créateur du fakemon.
      * @return DataResponse
      */
-    public function insertFakemon($fakemon): array
+    public function insertFakemon(array $fakemon, string $apiKey): array
     {
-        $sql = "INSERT INTO creature (nom,id_type1,id_type2,hp,atk,def,sp_atk,sp_def,speed,description) 
-        VALUES (:nom,:id_type1,:id_type2,:hp,:atk,:def,:sp_atk,:sp_def,:speed,:description);";
+        $idUsager = $this->chercherUsagerId($apiKey);
+        if (empty($idUsager)) {
+            $idUsager = [
+                "id" => 0
+            ];
+        }
+
+        $sql = "INSERT INTO creature (nom,id_type1,id_type2,hp,atk,def,sp_atk,sp_def,speed,description,id_usager) 
+        VALUES (:nom,:id_type1,:id_type2,:hp,:atk,:def,:sp_atk,:sp_def,:speed,:description,:id_usager);";
 
         $params = [
-            "nom"=> $fakemon["nom"] ?? "",
-            "id_type1"=> $fakemon["id_type1"] ?? 1,
-            "id_type2"=> $fakemon["id_type2"] ?? 1,
+            "nom" => $fakemon["nom"] ?? "",
+            "id_type1" => $fakemon["id_type1"] ?? 1,
+            "id_type2" => $fakemon["id_type2"] ?? 1,
             "hp" => $fakemon["hp"] ?? 0,
-            "atk"=> $fakemon["atk"] ?? 0,
-            "def"=> $fakemon["def"] ?? 0,
-            "sp_atk"=> $fakemon["sp_atk"] ?? 0,
-            "sp_def"=> $fakemon["sp_def"] ?? 0,
-            "speed"=> $fakemon["speed"] ?? 0,
-            "description"=> $fakemon["description"] ?? ""
+            "atk" => $fakemon["atk"] ?? 0,
+            "def" => $fakemon["def"] ?? 0,
+            "sp_atk" => $fakemon["sp_atk"] ?? 0,
+            "sp_def" => $fakemon["sp_def"] ?? 0,
+            "speed" => $fakemon["speed"] ?? 0,
+            "description" => $fakemon["description"] ?? "",
+            "id_usager" => $idUsager["id"]
         ];
 
         $query = $this->connection->prepare($sql);
@@ -103,45 +221,77 @@ class FakemonRepository
 
         $idFakemon = $this->connection->lastInsertId();
         $resultat = $this->afficherFakemonById($idFakemon);
-        
+
         return $resultat ?? [];
     }
 
+    /**
+     * Sélectionne l'id de l'usager avec son username.
+     * 
+     * @param string la clé d'api.
+     * @return DataResponse
+     */
+    public function chercherUsagerId(string $api_key): array
+    {
+        try {
+            $sql = "SELECT id FROM usager WHERE api_key = :api_key ;";
 
-     /**
+            $params = [
+                "api_key" => $api_key
+            ];
+
+            $query = $this->connection->prepare($sql);
+            $query->execute($params);
+
+            $result = $query->fetchAll(PDO::FETCH_ASSOC);
+
+
+
+            return $result[0] ?? [];
+        } catch (\Throwable $th) {
+            return []; 
+        }
+    }
+
+
+    /**
      * Permet de supprimer un fakemon.
      * 
      * @param int l'id du fakemon
      * @return DataResponse
      */
-    public function deleteFakemon($id): array
+    public function deleteFakemon(int $id,int $idUsager): array
     {
         $test = $this->afficherFakemonById($id);
 
-        if (empty($test)){
-           $resultat = [
-            "data"=> [],
-            "status"=>404
-           ];
-            
-            return $resultat ?? [];
-        }
-        else {
-            $sql = "DELETE FROM creature WHERE id=:id ;";
-    
-            $params = [
-                "id"=> $id
-            ];
-    
-            $query = $this->connection->prepare($sql);
-            $query->execute($params);
-
+        if (empty($test)) {
             $resultat = [
-                "data"=>$test,
-                "status"=> 200
+                "data" => [],
+                "status" => 404
             ];
-            
             return $resultat ?? [];
+        } else {
+            if($test["id_usager"] === $idUsager){
+                $sql = "DELETE FROM creature WHERE id=:id ;";
+
+                $params = [
+                    "id" => $id
+                ];
+
+                $query = $this->connection->prepare($sql);
+                $query->execute($params);
+
+                $resultat = [
+                    "data" => $test,
+                    "status" => 200
+                ];
+
+                return $resultat ?? [];
+            } 
+            else {
+                return [];
+            }
+            
         }
     }
 
@@ -150,74 +300,51 @@ class FakemonRepository
      * Permet de modifier un fakemon
      * 
      * @param array les informations du fakemon
-     * @param int l'id du fakemon
+     * @param string la clé d'api
      * @return DataResponse
      */
-    public function updateFakemon(array $fakemon,int $id): array
+    public function updateFakemon(array $fakemon, int $idFakemon, string $apikey): array
     {
 
-        $test = $this->afficherFakemonById($id);
+        $test = $this->afficherFakemonById($idFakemon);
 
-        if (empty($test)){
-            $sql = "INSERT INTO creature (nom,id_type1,id_type2,hp,atk,def,sp_atk,sp_def,speed,description) 
-            VALUES (:nom,:id_type1,:id_type2,:hp,:atk,:def,:sp_atk,:sp_def,:speed,:description);";
-    
-            $params = [
-                "nom"=> $fakemon["nom"] ?? "",
-                "id_type1"=> $fakemon["id_type1"] ?? 1,
-                "id_type2"=> $fakemon["id_type2"] ?? 1,
-                "hp" => $fakemon["hp"] ?? 0,
-                "atk"=> $fakemon["atk"] ?? 0,
-                "def"=> $fakemon["def"] ?? 0,
-                "sp_atk"=> $fakemon["sp_atk"] ?? 0,
-                "sp_def"=> $fakemon["sp_def"] ?? 0,
-                "speed"=> $fakemon["speed"] ?? 0,
-                "description"=> $fakemon["description"] ?? ""
-            ];
-    
-            $query = $this->connection->prepare($sql);
-            $query->execute($params);
-    
-            $result = $query->fetchAll(PDO::FETCH_ASSOC);
-    
-            $idFakemon = $this->connection->lastInsertId();
+        if (empty($test)) {
+            $resultat = $this->insertFakemon($fakemon, $apikey);
             $resultat = [
-                "data"=>$this->afficherFakemonById($idFakemon),
-                "status"=> 201
+                "data" => $resultat,
+                "status" => 201
             ];
-            
+
             return $resultat ?? [];
-        }
-        else {
+        } else {
             $sql = "UPDATE creature
              SET nom=:nom,id_type1=:id_type1,id_type2=:id_type2,hp=:hp,atk=:atk,def=:def,sp_atk=:sp_atk,sp_def=:sp_def,speed=:speed,description=:description
               WHERE id = :id";
-    
+
             $params = [
-                "id"=> $id,
-                "nom"=> $fakemon["nom"] ?? "",
-                "id_type1"=> $fakemon["id_type1"] ?? 1,
-                "id_type2"=> $fakemon["id_type2"] ?? 1,
+                "id" => $idFakemon,
+                "nom" => $fakemon["nom"] ?? "",
+                "id_type1" => $fakemon["id_type1"] ?? 1,
+                "id_type2" => $fakemon["id_type2"] ?? 1,
                 "hp" => $fakemon["hp"] ?? 0,
-                "atk"=> $fakemon["atk"] ?? 0,
-                "def"=> $fakemon["def"] ?? 0,
-                "sp_atk"=> $fakemon["sp_atk"] ?? 0,
-                "sp_def"=> $fakemon["sp_def"] ?? 0,
-                "speed"=> $fakemon["speed"] ?? 0,
-                "description"=> $fakemon["description"] ?? ""
+                "atk" => $fakemon["atk"] ?? 0,
+                "def" => $fakemon["def"] ?? 0,
+                "sp_atk" => $fakemon["sp_atk"] ?? 0,
+                "sp_def" => $fakemon["sp_def"] ?? 0,
+                "speed" => $fakemon["speed"] ?? 0,
+                "description" => $fakemon["description"] ?? ""
             ];
-    
+
             $query = $this->connection->prepare($sql);
             $query->execute($params);
 
             $resultat = [
-                "data"=>$this->afficherFakemonById($id),
-                "status"=> 200
+                "data" => $this->afficherFakemonById($idFakemon),
+                "status" => 200
             ];
-            
+
             return $resultat ?? [];
         }
-       
     }
 
     /**
@@ -225,18 +352,21 @@ class FakemonRepository
      * 
      * @return DataResponse
      */
-    public function selectFakemon(): array
+    public function selectFakemon(string $idUsager): array
     {
-        $sql = "SELECT * FROM creature ;";
+        $sql = "SELECT * FROM creature WHERE id_usager = :idUsager ;";
 
 
         $query = $this->connection->prepare($sql);
-        $query->execute();
+        $params = [
+            "idUsager" => $idUsager
+        ];
+        $query->execute($params);
 
         $result = $query->fetchAll(PDO::FETCH_ASSOC);
 
-        
-        
+
+
         return $result ?? [];
     }
 
@@ -256,10 +386,8 @@ class FakemonRepository
 
         $result = $query->fetchAll(PDO::FETCH_ASSOC);
 
-        
-        
+
+
         return $result ?? [];
     }
-
 }
-
